--- src/main/java/frc/robot/commands/SubsystemCommandFactory.java
@@ -1,20 +1,0 @@
-package frc.robot.commands;
-
-import edu.wpi.first.wpilibj2.command.Command;
-import edu.wpi.first.wpilibj2.command.Commands;
-import frc.robot.subsystems.subsystem.Subsystem;
-
-public class SubsystemCommandFactory {
-
-  private SubsystemCommandFactory() {}
-
-  /**
-   * Example method that creates an inline command for an example subsystem
-   *
-   * @param Subsystem subsystem
-   * @return Command (example inline runOnce command)
-   */
-  public Command getExampleSubsystemCommand(Subsystem subsystem) {
-    return Commands.runOnce(subsystem::exampleMethod, subsystem);
-  }
-}
