--- src/main/java/frc/robot/operator_interface/SimDualJoysticksOI.java
@@ -1,77 +1,0 @@
-// Copyright (c) FIRST and other WPILib contributors.
-// Open Source Software; you can modify and/or share it under the terms of
-// the WPILib BSD license file in the root directory of this project.
-
-package frc.robot.operator_interface;
-
-import edu.wpi.first.wpilibj2.command.button.CommandJoystick;
-import edu.wpi.first.wpilibj2.command.button.Trigger;
-
-/** Class for controlling the robot with two joysticks. */
-public class SimDualJoysticksOI implements OperatorInterface {
-  private final CommandJoystick translateJoystick;
-  private final CommandJoystick rotateJoystick;
-  private final Trigger[] translateJoystickButtons;
-  private final Trigger[] rotateJoystickButtons;
-
-  public SimDualJoysticksOI(int translatePort, int rotatePort) {
-    translateJoystick = new CommandJoystick(translatePort);
-    rotateJoystick = new CommandJoystick(rotatePort);
-
-    // buttons use 1-based indexing such that the index matches the button number; leave index 0 set
-    // to null
-    this.translateJoystickButtons = new Trigger[13];
-    this.rotateJoystickButtons = new Trigger[13];
-
-    for (int i = 1; i < translateJoystickButtons.length; i++) {
-      translateJoystickButtons[i] = translateJoystick.button(i);
-      rotateJoystickButtons[i] = rotateJoystick.button(i);
-    }
-  }
-
-  @Override
-  public double getTranslateX() {
-    return -translateJoystick.getY();
-  }
-
-  @Override
-  public double getTranslateY() {
-    return -translateJoystick.getX();
-  }
-
-  @Override
-  public double getRotate() {
-    return -rotateJoystick.getX();
-  }
-
-  @Override
-  public Trigger getFieldRelativeButton() {
-    return translateJoystickButtons[3];
-  }
-
-  @Override
-  public Trigger getResetGyroButton() {
-    return rotateJoystickButtons[3];
-  }
-
-  @Override
-  public Trigger getLock180Button() {
-    return new Trigger(() -> false);
-  }
-
-  @Override
-  public Trigger getXStanceButton() {
-    return rotateJoystickButtons[4];
-  }
-
-  @Override
-  public Trigger getVisionIsEnabledSwitch() {
-    // vision is always enabled with dual joysticks as there is no switch to disable
-    return new Trigger(() -> true);
-  }
-
-  @Override
-  public Trigger getResetPoseToVisionButton() {
-    return translateJoystickButtons[1];
-  }
-}
